src/requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dnspython==2.4.2
gogettr==0.8.0
idna==3.4
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
mysql-connector-python==8.1.0
numpy==1.26.0
packaging==23.1
pandas==2.1.1
pluggy==1.3.0
protobuf==4.21.12
psycopg2-binary==2.9.8
Pygments==2.16.1
pytest==7.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tzdata==2023.3
urllib3==2.0.5
